+++ SSC-C/manifest.txt
@@ -1,2 +1,2 @@
-Main-Class: cz.muni.fi.sscc.Main
+Main-Class: sscc.cz.mipit.sscc.Main
 Class-Path: antlr-4.13.2-complete.jar
